requirements.txt
celery==4.4.7
django-filter==2.2.0
django==3.0.5
djangorestframework==3.11.0
flake8==3.7.9
gunicorn==20.0.4
psycopg2-binary==2.8.6
boto3==1.15.3
django-storages==1.10.1
Pillow==7.2.0
pandas
openpyxl
